OAOTC-759 Refactored by adding to transfer details directory
