+++ app/controllers/transferDetails/AmountOfTaxDeductedController.scala
@@ -14,31 +14,31 @@
  * limitations under the License.
  */
 
-package controllers
+package controllers.transferDetails
 
 import controllers.actions._
-import forms.CashAmountInTransferFormProvider
-import javax.inject.Inject
+import forms.AmountOfTaxDeductedFormProvider
 import models.Mode
-import pages.CashAmountInTransferPage
+import pages.transferDetails.AmountOfTaxDeductedPage
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import repositories.SessionRepository
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendBaseController
-import views.html.CashAmountInTransferView
+import views.html.transferDetails.AmountOfTaxDeductedView
 
+import javax.inject.Inject
 import scala.concurrent.{ExecutionContext, Future}
 
-class CashAmountInTransferController @Inject() (
+class AmountOfTaxDeductedController @Inject() (
     override val messagesApi: MessagesApi,
     sessionRepository: SessionRepository,
     identify: IdentifierAction,
     getData: DataRetrievalAction,
     requireData: DataRequiredAction,
     displayData: DisplayAction,
-    formProvider: CashAmountInTransferFormProvider,
+    formProvider: AmountOfTaxDeductedFormProvider,
     val controllerComponents: MessagesControllerComponents,
-    view: CashAmountInTransferView
+    view: AmountOfTaxDeductedView
   )(implicit ec: ExecutionContext
   ) extends FrontendBaseController with I18nSupport {
 
@@ -46,7 +46,7 @@ class CashAmountInTransferController @Inject() (
 
   def onPageLoad(mode: Mode): Action[AnyContent] = (identify andThen getData andThen requireData andThen displayData) {
     implicit request =>
-      val preparedForm = request.userAnswers.get(CashAmountInTransferPage) match {
+      val preparedForm = request.userAnswers.get(AmountOfTaxDeductedPage) match {
         case None        => form
         case Some(value) => form.fill(value)
       }
@@ -61,9 +61,9 @@ class CashAmountInTransferController @Inject() (
           Future.successful(BadRequest(view(formWithErrors, mode))),
         value =>
           for {
-            updatedAnswers <- Future.fromTry(request.userAnswers.set(CashAmountInTransferPage, value))
+            updatedAnswers <- Future.fromTry(request.userAnswers.set(AmountOfTaxDeductedPage, value))
             _              <- sessionRepository.set(updatedAnswers)
-          } yield Redirect(CashAmountInTransferPage.nextPage(mode, updatedAnswers))
+          } yield Redirect(AmountOfTaxDeductedPage.nextPage(mode, updatedAnswers))
       )
   }
 }
