+++ app/views/transferDetails/AmountOfTaxDeductedView.scala.html
@@ -17,7 +17,6 @@
 @import viewmodels.InputWidth._
 @import uk.gov.hmrc.govukfrontend.views.html.components.FormWithCSRF
 @import uk.gov.hmrc.govukfrontend.views.html.components.GovukErrorSummary
-
 @import uk.gov.hmrc.govukfrontend.views.html.components.GovukButton
 @import views.ViewUtils.title
 @import viewmodels.govuk.all._
@@ -40,24 +39,23 @@
 
 @(form: Form[_], mode: Mode)(implicit request: DisplayRequest[_], messages: Messages)
 
-@layout(pageTitle = title(form, messages("cashAmountInTransfer.title"))) {
+@layout(pageTitle = title(form, messages("amountOfTaxDeducted.title"))) {
 
-    @formHelper(action = routes.CashAmountInTransferController.onSubmit(mode), Symbol("autoComplete") -> "on") {
+    @formHelper(action = controllers.transferDetails.routes.AmountOfTaxDeductedController.onSubmit(mode), Symbol("autoComplete") -> "on") {
 
         @if(form.errors.nonEmpty) {
             @govukErrorSummary(ErrorSummaryViewModel(form))
         }
 
         @qtNumber()
-
         @govukInput(
             InputViewModel(
-                field = form("cashInTransfer"),
-                label = LabelViewModel(Text(messages("cashAmountInTransfer.heading")))
+                field = form("taxDeducted"),
+                label = LabelViewModel(Text(messages("amountOfTaxDeducted.heading")))
                         .withCssClass("govuk-!-margin-bottom-6")
                         .asPageHeading(size = Large)
             )
-            .withHint(HintViewModel(messages("cashAmountInTransfer.hint")))
+            .withHint(HintViewModel(messages("amountOfTaxDeducted.hint")))
             .withPrefix(PrefixOrSuffix(content = Text("£")))
             .withWidth(Fixed5)
         )
